refactor: :ambulance: remove the function that users need to manually set location when adding a report
