+++ src/locale/en-us/drawer.json
@@ -26,9 +26,6 @@
       "text": {
         "setLocation": "Location"
       },
-      "button": {
-        "setLocation": "Mark"
-      },
       "select": {
         "setFloor": {
           "label": "Floor"
